# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Apple Connector API",
    "description": "Hybrid HTTP API for Messages.app, Reminders.app, Notes.app, Calendar.app, and Contacts.\n\n**Reads** use live SQLite connections (requires Full Disk Access).\n\n**Writes** for Reminders, Calendar events, and Contacts use EventKit / Contacts framework (requires Reminders, Calendars, and Contacts permissions in System Settings). Unsupported fields return `422`; smart lists, read-only calendars/containers return `403`. Post-save responses hydrate from SQLite with up to 5×100ms retries; `sync_pending: true` indicates the SQLite read path has not caught up yet.\n\nPagination uses keyset cursors only (default limit 50, maximum 200, newest first). Offsets are not supported.\n\nAuthentication, TLS, and network exposure controls are expected to be enforced by an external reverse proxy or firewall. This service does not implement authentication or TLS.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as this service (use the host you opened in the browser, e.g. http://127.0.0.1:3000 or http://localhost:3000)"
    }
  ],
  "paths": {
...
    "/openapi.json": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Get OpenAPI specification",
        "description": "Returns the OpenAPI 3.1 contract served by this API.",
        "operationId": "getOpenApiSpec",
        "responses": {
          "200": {
            "description": "OpenAPI 3.1 specification",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/calendar-accounts": {
      "get": {
        "tags": [
          "calendar-accounts"
        ],
        "summary": "List calendar accounts",
        "operationId": "listCalendarAccounts",
        "responses": {
          "200": {
            "description": "Calendar accounts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarAccountPageDto"
                }
              }
            }
          },
          "503": {
            "description": "Calendar database is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/calendars": {
      "get": {
        "tags": [
          "calendars"
        ],
        "summary": "List calendars",
        "operationId": "listCalendars",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return. Defaults to 50. Keyset pagination only; no offsets.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 200,
              "minimum": 1
            },
            "style": "form",
            "example": 50
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "URL-safe versioned cursor for the next page. Results are ordered newest first.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "style": "form",
            "example": "v1.eyJkYXRlIjoxNzA0MDk2MDAwfQ"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated calendars",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarPageDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Calendar database is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/chats": {
      "get": {
        "tags": [
          "chats"
        ],
        "summary": "List chats",
        "description": "Returns chats ordered by recent activity. Uses keyset pagination (newest first).",
        "operationId": "listChats",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return. Defaults to 50. Keyset pagination only; no offsets.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 200,
              "minimum": 1
            },
            "style": "form",
            "example": 50
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "URL-safe versioned cursor for the next page. Results are ordered newest first.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "style": "form",
            "example": "v1.eyJkYXRlIjoxNzA0MDk2MDAwfQ"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated chat summaries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatPageDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Messages database is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/contacts": {
      "get": {
        "tags": [
          "contacts"
        ],
        "summary": "List contacts globally as JSON",
        "operationId": "listContacts",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 200,
              "minimum": 1
            },
            "style": "form",
            "example": 50
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "style": "form",
            "example": "v1.eyJyb3dfaWQiOjF9"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 256
            },
            "style": "form",
            "example": "jane"
          },
          {
            "name": "container_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "style": "form"
          },
          {
            "name": "group_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated contacts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactPageDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Contacts databases are unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/contacts/carddav": {
      "get": {
        "tags": [
          "contacts"
        ],
        "summary": "List contacts globally as CardDAV XML",
        "operationId": "listContactsCarddav",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 200,
              "minimum": 1
            },
            "style": "form",
            "example": 50
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "style": "form",
            "example": "v1.eyJyb3dfaWQiOjF9"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 256
            },
            "style": "form",
            "example": "jane"
          },
          {
            "name": "container_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "style": "form"
          },
          {
            "name": "group_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "description": "CardDAV multistatus",
            "content": {
              "application/carddav+xml": {}
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Contacts databases are unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/contacts/search": {
      "get": {
        "tags": [
          "contacts"
        ],
        "summary": "Search contacts",
        "operationId": "searchContacts",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 200,
              "minimum": 1
            },
            "style": "form",
            "example": 50
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "style": "form",
            "example": "v1.eyJyb3dfaWQiOjF9"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 256
            },
            "style": "form",
            "example": "jane"
          },
          {
            "name": "container_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "style": "form"
          },
          {
            "name": "group_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "description": "Matching contacts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactPageDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Contacts databases are unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/contacts/vcard": {
      "get": {
        "tags": [
          "contacts"
        ],
        "summary": "List contacts globally as vCard",
        "operationId": "listContactsVcard",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 200,
              "minimum": 1
            },
            "style": "form",
            "example": 50
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "style": "form",
            "example": "v1.eyJyb3dfaWQiOjF9"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 256
            },
            "style": "form",
            "example": "jane"
          },
          {
            "name": "container_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "style": "form"
          },
          {
            "name": "group_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "description": "vCard feed",
            "content": {
              "text/vcard": {}
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Contacts databases are unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/containers": {
      "get": {
        "tags": [
          "containers"
        ],
        "summary": "List contact containers",
        "operationId": "listContainers",
        "responses": {
          "200": {
            "description": "Contact containers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContainerPageDto"
                }
              }
            }
          },
          "503": {
            "description": "Contacts databases are unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CalendarAccountPageDto": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CalendarAccountDto"
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
    },
    "securitySchemes": {
      "reverse_proxy": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Authentication is expected to be enforced by an external reverse proxy or firewall. This API does not implement authentication or TLS."
      }
    }
  },
  "security": [
    {
      "reverse_proxy": []
    }
  ],
...
}